Update `term` to 0.6 (BREAKING CHANGE)
This also includes an update to `Color`, changing `Custom(u16)` to
`Custom(u32)`.

+++ src/lib.rs
@@ -312,7 +312,7 @@ pub enum Color {
     BrightMagenta,
     BrightCyan,
     BrightWhite,
-    Custom(u16),
+    Custom(u32),
 }
 
 impl Color {
